Refactor churn rate analysis.
Instead of simply determining the amount of joining and leaving relays,
we now use the formula in Section 2.1 of
<http://www.cs.berkeley.edu/~istoica/papers/2006/churn.pdf>.

+++ similarity.go
@@ -41,7 +41,7 @@ type DescriptorSimilarity struct {
 	ORPortDiff      uint16
 	SharedFprPrefix uint32
 	LevenshteinDist int
-	SimilarityScore int
+	SimilarityScore float64
 
 	SameFamily   bool
 	SameAddress  bool
@@ -112,7 +112,7 @@ func (s *DescriptorSimilarity) genStringSimilarity() {
 			s.desc1.ORPort, s.desc2.ORPort)
 	}
 
-	s.SimilarityScore = similarities
+	s.SimilarityScore = float64(similarities)
 	s.StringSummary = fmt.Sprintf("%d similarities%s:\n"+
 		"%s%s%s%s%s%s%s%s",
 		similarities, family,
